refactor name + change css
