+++ src/components/StatusBadge/index.tsx
@@ -2,5 +2,5 @@ import styled from "@emotion/styled";
 import Badge from "@leafygreen-ui/badge";
 
 export const StatusBadge = styled(Badge)`
-  text-transform: lowercase;
+  text-transform: capitalize;
 `;
